Created component FavoriteButton (#12)
