+++ Mobile/src/Component/FavoriteButton/styles.js
@@ -1,0 +1,20 @@
+import {StyleSheet} from 'react-native';
+
+const styles = StyleSheet.create({
+  favoriteButtonView: {
+    justifyContent: 'center',
+    alignItems: 'center',
+    borderRadius: 100,
+
+    position: 'absolute',
+    right: 1,
+  },
+  favoriteButton: {
+    justifyContent: 'center',
+    alignItems: 'center',
+    borderRadius: 100,
+    padding: 5,
+    alignSelf: 'center',
+  },
+});
+export default styles;
