+++ Mobile/src/Component/ImageSwipe/index.js
@@ -8,15 +8,13 @@ import {
   useWindowDimensions,
 } from 'react-native';
 
-import {RectButton} from 'react-native-gesture-handler';
-import Ionicon from 'react-native-vector-icons/Ionicons';
 import styles from './styles';
+import FavoriteButton from '../FavoriteButton';
 export default function ImageSwipe({img}) {
   const width = useWindowDimensions().width;
   const height = width * 0.6;
 
   const [active, setActive] = useState(0);
-  const [isFavorite, setIsFavorite] = useState(false);
 
   const change = ({nativeEvent}) => {
     const slide = Math.ceil(
@@ -27,9 +25,6 @@ export default function ImageSwipe({img}) {
     }
   };
 
-  function toggleFavorite() {
-    setIsFavorite(!isFavorite);
-  }
   return (
     <View>
       <ScrollView
@@ -46,18 +41,7 @@ export default function ImageSwipe({img}) {
           />
         ))}
       </ScrollView>
-      <View style={styles.favoriteButtonView}>
-        <RectButton
-          style={styles.favoriteButton}
-          onPress={() => {
-            toggleFavorite();
-          }}>
-          {isFavorite == true 
-            ? (<Ionicon name={'ios-heart'} size={50} color={'#CB2F2F'} />)
-            : (<Ionicon name={'ios-heart-outline'} size={50} color={'#CB2F2F'} />)
-          }
-        </RectButton>
-      </View>
+      <FavoriteButton />
       <View style={styles.pagination}>
         {img.map((i, k) => (
           <Text key={k} style={k === active ? styles.activeDot : styles.dot}>
